+++ src/utils/__tests__/impressumParser.test.ts
@@ -1,4 +1,7 @@
 
+/**
+ * @jest-environment jsdom
+ */
 import { parseImpressum } from '../impressumParser';
 
 const minimalImpressumHtml = `